+++ .scannerwork/report-task.txt
@@ -1,0 +1,5 @@
+projectKey=AccountApp
+serverUrl=http://localhost:9000
+dashboardUrl=http://localhost:9000/dashboard/index/AccountApp
+ceTaskId=AWbFcecsThYH9a9oqJr9
+ceTaskUrl=http://localhost:9000/api/ce/task?id=AWbFcecsThYH9a9oqJr9
