--- src/com/qa/account/AccountApplication/AccountInterface.java
@@ -1,8 +1,0 @@
-package com.qa.account.AccountApplication;
-
-public interface AccountInterface {
-
-	void addAccount(String firstName, String lastName, String accountNumber);
-	
-	Account retrieveAccount(String accountNumber);
-}
