--- src/com/qa/account/AccountApplication/App.java
@@ -1,13 +1,0 @@
-package com.qa.account.AccountApplication;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
